Fix the warnings.

+++ EasyPlayer/NJAudioEngine.h
@@ -26,5 +26,5 @@
 - (void)start;
 - (void)pause;
 - (void)stop;
-- (void)setVolume:(CGFloat)inVolume forBusIndex:(NSUInteger)inBusIndex;
+- (void)setVolume:(CGFloat)inVolume forBusIndex:(UInt32)inBusIndex;
 @end
